backend/requirements.txt
fastapi>=0.109.0,<0.112.0
uvicorn[standard]>=0.27.0,<0.30.0
python-multipart>=0.0.6,<0.1.0
# paper-qa и его зависимости:
paper-qa>=5.0.0,<6.0.0
litellm>=1.44.0,<2.0.0
pydantic>=2.10.1,<2.11.0
pydantic-settings>=2.1.0,<3.0.0
ollama>=0.1.7,<0.2.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
bcrypt>=4.0.0,<4.2.0
aiofiles>=23.2.1,<24.0.0
httpx>=0.27.0,<0.28.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
